+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Updated] - 2025-12-02
+
+- Updated extension icon
+- Added persistence for last used format and compression settings
+- Added support for RAW camera formats (CR3, CR2, DNG, NEF, ARW, RAF, ORF, RW2)
+
 ## [Initial Version] - 2024-11-29
 
 - Initial release of EXR Converter